pyproject.toml
[project]
name = "crypto-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "openai-agents>=0.6.7",
    "uvicorn>=0.40.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
]
